Implement tasks of 9 lesson
