+++ src/main/java/handlers/HistoryHandler.java
@@ -7,8 +7,9 @@ import java.io.IOException;
 
 public class HistoryHandler extends BaseHttpHandler implements HttpHandler {
 
+    @Override
     public void handle(HttpExchange exchange) throws IOException {
-        writeGetListResponse(exchange, fileBackedTaskManager.getHistory(), 200);
+        super.handle(exchange);
     }
 }
