+++ src/main/java/handlers/PrioritizedHandler.java
@@ -7,7 +7,8 @@ import java.io.IOException;
 
 public class PrioritizedHandler extends BaseHttpHandler implements HttpHandler {
 
+    @Override
     public void handle(HttpExchange exchange) throws IOException {
-        writeGetListResponse(exchange, fileBackedTaskManager.getPrioritizedTasks(), 200);
+        super.handle(exchange);
     }
 }
